nice refactor, it works!
