+++ packages/studio/src/previewEntry.tsx
@@ -1,42 +1,33 @@
+import React from 'react';
 import type {render} from 'react-dom';
-import {createPortal} from 'react-dom';
 import ReactDOM from 'react-dom/client';
 import {Internals} from 'remotion';
 import '../styles/styles.css';
-import {Editor} from './components/Editor';
-import {EditorContexts} from './components/EditorContexts';
-import {ServerDisconnected} from './components/Notifications/ServerDisconnected';
+import {NoRegisterRoot} from './components/NoRegisterRoot';
+import {Studio} from './components/Studio';
 import {openEventSource} from './helpers/event-source';
 
 Internals.CSSUtils.injectCSS(
 	Internals.CSSUtils.makeDefaultCSS(null, '#1f2428'),
 );
 
-const getServerDisconnectedDomElement = () => {
-	return document.getElementById('server-disconnected-overlay');
+const renderToDOM = (content: React.ReactElement) => {
+	if (ReactDOM.createRoot) {
+		ReactDOM.createRoot(Internals.getPreviewDomElement() as HTMLElement).render(
+			content,
+		);
+	} else {
+		(ReactDOM as unknown as {render: typeof render}).render(
+			content,
+			Internals.getPreviewDomElement(),
+		);
+	}
 };
 
-const content = (
-	<Internals.RemotionRoot numberOfAudioTags={window.remotion_numberOfAudioTags}>
-		<EditorContexts>
-			<Editor />
-			{createPortal(
-				<ServerDisconnected />,
-				getServerDisconnectedDomElement() as HTMLElement,
-			)}
-		</EditorContexts>
-	</Internals.RemotionRoot>
-);
+renderToDOM(<NoRegisterRoot />);
 
-if (ReactDOM.createRoot) {
-	ReactDOM.createRoot(Internals.getPreviewDomElement() as HTMLElement).render(
-		content,
-	);
-} else {
-	(ReactDOM as unknown as {render: typeof render}).render(
-		content,
-		Internals.getPreviewDomElement(),
-	);
-}
+Internals.waitForRoot((NewRoot) => {
+	renderToDOM(<Studio rootComponent={NewRoot} />);
+});
 
 openEventSource();
